+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-24T14:42:39.367Z",
+    "2026-06-24T15:35:00.800Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {