+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3861,7 +3861,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KVX1B43WBX0EWTSTP37VZH8Z"),
+  id: z.string().default("01KVX4B5QRQGFJ9D6WF0Q0HZB0"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),