+++ packages/orq-rc/RELEASES.md
@@ -10249,3 +10249,13 @@ Based on:
 - [typescript v4.11.0-rc.46] packages/orq-rc
 ### Releases
 - [NPM v4.11.0-rc.46] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.46 - packages/orq-rc
+
+## 2026-06-24 15:37:55
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.785.0 (2.912.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.11.0-rc.47] packages/orq-rc
+### Releases
+- [NPM v4.11.0-rc.47] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.47 - packages/orq-rc